move deepeval to providers
